# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "家庭记账本 API",
    "description": "家庭共享记账 H5 应用后端 API 接口文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码，0 表示成功"
          },
          "data": {
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "提示信息"
          }
        },
        "required": ["code", "message"]
      },
...
      "FamilyMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "description": "1: 支出, 2: 收入"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
      "Record": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "amount": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "description": "1: 支出, 2: 收入"
          },
          "remark": {
            "type": "string",
            "nullable": true
          },
          "recordDate": {
            "type": "string",
            "format": "date"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "user": {
            "$ref": "#/components/schemas/FamilyMember"
          }
        }
      },
      "OverviewStats": {
        "type": "object",
        "properties": {
          "income": {
            "type": "string"
          },
          "expense": {
            "type": "string"
          },
          "balance": {
            "type": "string"
          }
        }
      },
      "CategoryStats": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer"
          },
          "categoryName": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "percentage": {
            "type": "integer"
          }
        }
      },
      "TrendStats": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string"
          },
          "income": {
            "type": "string"
          },
          "expense": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/record": {
      "get": {
        "tags": ["记账"],
        "summary": "获取账目列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "月份，格式：2024-01"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "页码，默认1"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "每页数量，默认20"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "类型：1支出 2收入"
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "分类ID"
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Record"
                              }
                            },
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "size": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["记账"],
        "summary": "创建账目",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "type", "categoryId", "recordDate"],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "金额"
                  },
                  "type": {
                    "type": "integer",
                    "description": "类型：1支出 2收入"
                  },
                  "categoryId": {
                    "type": "integer",
                    "description": "分类ID"
                  },
                  "remark": {
                    "type": "string",
                    "description": "备注"
                  },
                  "recordDate": {
                    "type": "string",
                    "format": "date",
                    "description": "记账日期"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "记账成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Record"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/record/{id}": {
      "get": {
        "tags": ["记账"],
        "summary": "获取账目详情",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Record"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["记账"],
        "summary": "更新账目",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "type": {
                    "type": "integer"
                  },
                  "categoryId": {
                    "type": "integer"
                  },
                  "remark": {
                    "type": "string"
                  },
                  "recordDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Record"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["记账"],
        "summary": "删除账目",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/overview": {
      "get": {
        "tags": ["统计"],
        "summary": "收支概览",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "月份，格式：2024-01"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OverviewStats"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/category": {
      "get": {
        "tags": ["统计"],
        "summary": "分类统计",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "月份，格式：2024-01"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "类型：1支出 2收入，默认1"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CategoryStats"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/trend": {
      "get": {
        "tags": ["统计"],
        "summary": "趋势图表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "查询月份数，默认6"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TrendStats"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/member": {
      "get": {
        "tags": ["统计"],
        "summary": "成员贡献排行",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "月份，格式：2024-01"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "userId": {
                                "type": "integer"
                              },
                              "nickname": {
                                "type": "string"
                              },
                              "avatar": {
                                "type": "string",
                                "nullable": true
                              },
                              "totalAmount": {
                                "type": "string"
                              },
                              "recordCount": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}